+++ test/screenshot/drawer/dismissible.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import DrawerTest from './DrawerTest';
+
+export default () => {
+  return <DrawerTest dismissible title='Dismissible Drawer' />
+};
