+++ test/screenshot/drawer/modal.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import DrawerTest from './DrawerTest';
+
+export default () => {
+  return <DrawerTest modal title='Modal Drawer' />
+};
