+++ test/screenshot/drawer/permanent.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import DrawerTest from './DrawerTest';
+
+export default () => {
+  return <DrawerTest hideNavigationIcon title='Permanent Drawer' />
+};
