+++ test/screenshot/drawer/variants.js
@@ -1,0 +1,5 @@
+export default [
+  'permanent',
+  'dismissible',
+  'modal',
+];
