+++ test/screenshot/index.js
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 import {HashRouter, Route} from 'react-router-dom';
 import App from './App';
 import topAppBarVariants from './top-app-bar/variants';
+import drawerVariants from './drawer/variants';
 
 import './index.scss';
 
@@ -14,6 +15,13 @@ ReactDOM.render((
         const Component = require(`./${componentPath}/index.js`).default;
         return (
           <Route exact key={componentPath} path={`/${componentPath}`} component={Component}/>
+        );
+      })}
+      {drawerVariants.map((variant) => {
+        const path = `drawer/${variant}`;
+        const Component = require(`./drawer/${variant}.js`).default;
+        return (
+          <Route exact key={path} path={`/${path}`} component={Component}/>
         );
       })}
       {topAppBarVariants.map((variant) => {
